requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
beautifulsoup4
jinja2
pydantic 